fix : 거리 재계산

+++ backend/src/main/java/com/ssafy/backend/dto/HospitalInfoDto.java
@@ -29,7 +29,7 @@ public class HospitalInfoDto {
 
 	private Integer minPrice;
 
-	private Long distance;
+	private Integer distance;
 
 	private String treatmentName;
 }
