+++ backend/src/main/java/com/ssafy/backend/service/HospitalService.java
@@ -56,7 +56,7 @@ public class HospitalService {
 			double lat = price.getHospital().getLatitude();
 			// 경도
 			double lon = price.getHospital().getLongitude();
-			Long dis = locationDistance(userLatitude, userLongitude, lat, lon);
+			int dis = locationDistance(userLatitude, userLongitude, lat, lon);
 
 			// 거리 저장
 			if (dis <= disLimit) {
@@ -91,7 +91,7 @@ public class HospitalService {
 			double lat = price.getHospital().getLatitude();
 			// 경도
 			double lon = price.getHospital().getLongitude();
-			Long dis = locationDistance(userLatitude, userLongitude, lat, lon);
+			int dis = locationDistance(userLatitude, userLongitude, lat, lon);
 
 			// 거리 저장
 			if (dis <= disLimit) {
@@ -114,7 +114,7 @@ public class HospitalService {
 
 	}
 
-	private static Long locationDistance(double lat1, double lon1, double lat2, double lon2) {
+	private static int locationDistance(double lat1, double lon1, double lat2, double lon2) {
 		double theta = lon1 - lon2;
 		double dist =
 			Math.sin(deg2rad(lat1)) * Math.sin(deg2rad(lat2)) + Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2))
@@ -123,7 +123,7 @@ public class HospitalService {
 		dist = rad2deg(dist);
 		dist = dist * 60 * 1.1515 * 1609.344;
 
-		return (long) dist; //단위 meter
+		return (int) dist; //단위 meter
 	}
 
 	//10진수를 radian(라디안)으로 변환
